# Conflicts:
#	pom.xml
#	src/main/java/app/RateMyCampusAppBuilder.java
#	src/main/java/data_access/DBReviewAccessObject.java
#	src/main/java/entity/UserReview.java
#	src/main/java/interface_adapters/create_review/CreateReviewPresenter.java

+++ src/main/java/entity/UserReviewFactory.java
@@ -8,5 +8,5 @@ public interface UserReviewFactory {
      * @param comment the comment provided in the review
      * @return the new user review
      */
-    UserReview create(User user, double rating, String comment);
+    UserReview create(User user, int rating, String comment);
 }
